refactor: converted login page to SSR
