+++ src/app/auth/login/page.js
@@ -1,9 +1,3 @@
-"use client";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { useToast } from "@/hooks/use-toast";
 import {
   Card,
   CardContent,
@@ -11,47 +5,9 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { useAuth } from "@/lib/context/Auth.context";
-import Link from "next/link";
+import LoginForm from "@/components/pages/LoginForm";
 
-const loginSchema = z.object({
-  username: z.string().min(2, {
-    message: "Username must be at least 2 characters.",
-  }),
-  password: z.string().min(6, {
-    message: "Password must be at least 6 characters.",
-  }),
-});
-
-export default function LoginPage() {
-  const { toast } = useToast();
-
-  const { login } = useAuth();
-
-  const form = useForm({
-    resolver: zodResolver(loginSchema),
-    defaultValues: {
-      username: "",
-      password: "",
-    },
-  });
-
-  const onSubmit = async (data) => {
-    toast({ description: "Logging In... 🫸🏻" });
-
-    await login(data);
-  };
-
+const LoginPage = () => {
   return (
     <div className="flex justify-center items-center h-[95vh]">
       <Card className="w-[350px]">
@@ -62,54 +18,11 @@ export default function LoginPage() {
           </CardDescription>
         </CardHeader>
         <CardContent>
-          <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-              <FormField
-                control={form.control}
-                name="username"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Username</FormLabel>
-                    <FormControl>
-                      <Input placeholder="Enter your username" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="password"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Password</FormLabel>
-                    <FormControl>
-                      <Input
-                        type="password"
-                        placeholder="Enter your password"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <div className="flex flex-col space-y-2">
-                <Button type="submit" className="flex justify-center">
-                  Login
-                </Button>
-                <p className="mt-4 text-xs text-center text-balance">
-                  Don&apos;t have an account?{" "}
-                  <Link href="/auth/register" className="underline">
-                    Sign Up
-                  </Link>
-                </p>
-              </div>
-            </form>
-          </Form>
+          <LoginForm />
         </CardContent>
       </Card>
     </div>
   );
-}
+};
+
+export default LoginPage;
