+++ src/components/pages/LoginForm.jsx
@@ -1,0 +1,90 @@
+"use client";
+
+import React from "react";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import loginSchema from "@/utils/validation/loginSchema";
+import { useToast } from "@/hooks/use-toast";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+import { useAuth } from "@/lib/context/Auth.context";
+import { Button } from "@/components/ui/button";
+import Link from "next/link";
+
+const LoginForm = () => {
+  const { toast } = useToast();
+
+  const { login } = useAuth();
+
+  const form = useForm({
+    resolver: zodResolver(loginSchema),
+    defaultValues: {
+      username: "",
+      password: "",
+    },
+  });
+
+  const onSubmit = async (data) => {
+    toast({ description: "Logging In... 🫸🏻" });
+
+    await login(data);
+  };
+
+  return (
+    <Form {...form}>
+      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
+        <FormField
+          control={form.control}
+          name="username"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Username</FormLabel>
+              <FormControl>
+                <Input placeholder="Enter your username" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="password"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Password</FormLabel>
+              <FormControl>
+                <Input
+                  type="password"
+                  placeholder="Enter your password"
+                  {...field}
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <div className="flex flex-col space-y-2">
+          <Button type="submit" className="flex justify-center">
+            Login
+          </Button>
+          <p className="mt-4 text-xs text-center text-balance">
+            Don&apos;t have an account?{" "}
+            <Link href="/auth/register" className="underline">
+              Sign Up
+            </Link>
+          </p>
+        </div>
+      </form>
+    </Form>
+  );
+};
+
+export default LoginForm;
